# Fix Linux V4L2 latency - reduce from 12 to 8 frames