--- a/src/common/version.h
+++ b/src/common/version.h
@@ -1,10 +1,10 @@
 #pragma once
 
 #define VERSION_MAJOR 1
-#define VERSION_MINOR 6
-#define VERSION_PATCH 7
+#define VERSION_MINOR 7
+#define VERSION_PATCH 0
 
-#define VERSION_STRING "1.6.7"
+#define VERSION_STRING "1.7.0"
 
 // Additional defines for main.cpp compatibility
 #define NDI_BRIDGE_VERSION VERSION_STRING