## Overview
This PR implements comprehensive latency optimizations for Linux V4L2 capture to match the Windows Media Foundation performance of 8 frames latency.

## Current Situation
- Windows Media Foundation: 8 frames latency ✅
- Linux V4L2: 12 frames latency ❌

## Changes Implemented
1. **Remove sleep delays in multi-threaded pipeline**
   - Removed 100μs sleeps in convert and send threads
   - Use tight polling loops instead

2. **Reduce buffer counts and queue depths**
   - V4L2 buffer count: 10 → 6 (normal), 4 (low latency)
   - Capture queue depth: 5 → 3 (normal), 2 (low latency)
   - Convert queue depth: 5 → 2 (normal), 1 (low latency)

3. **Add single-threaded capture mode**
   - New command-line option: `--single-thread`
   - Eliminates inter-thread queues
   - Similar to Windows Media Foundation implementation

4. **Add low latency mode**
   - New command-line option: `--low-latency`
   - Forces single-threaded operation
   - Uses minimal buffers and immediate polling
   - Most aggressive settings for lowest latency

5. **Optimize polling timeouts**
   - Multi-threaded: 1ms → 0ms (immediate)
   - Single-threaded: 5ms → 1ms
   - Low latency: always 0ms

6. **Add latency monitoring**
   - End-to-end latency tracking
   - Frame-to-frame latency measurement
   - Detailed statistics logging

## Command-Line Options
```bash
# Standard multi-threaded (improved)
./ndi-bridge -d /dev/video0 -n "Stream"

# Single-threaded mode (lower latency)
./ndi-bridge -d /dev/video0 -n "Stream" --single-thread

# Low latency mode (lowest latency)
./ndi-bridge -d /dev/video0 -n "Stream" --low-latency

# With statistics
./ndi-bridge -d /dev/video0 -n "Stream" --single-thread -v
```

## Testing Required
- [ ] Test with 60fps camera
- [ ] Measure round-trip latency
- [ ] Compare single vs multi-threaded modes
- [ ] Verify CPU usage remains reasonable
- [ ] Test with different camera formats (YUYV, MJPEG, etc.)

## Expected Results
- Multi-threaded mode: ~10 frames (improved from 12)
- Single-threaded mode: ~8 frames (matches Windows)
- Low latency mode: ~8 frames (most aggressive)

## Related Issues
- Follows successful Windows latency fix from PR #13
- Implements recommendations from docs/LINUX_LATENCY_INVESTIGATION.md

## Files Changed
- `src/common/version.h` - Bumped to v1.7.0
- `src/linux/v4l2/v4l2_capture.h` - Added low latency support
- `src/linux/v4l2/v4l2_capture.cpp` - Implemented all optimizations
- `src/main.cpp` - Added new command-line options
- `CHANGELOG.md` - Documented changes